# sys_know/generated/openapi/schema.yaml
openapi: 3.0.3
info:
  title: TalentOS API
  version: 1.0.0
  description: Auto-generated API schema for testing and documentation
paths:
  /api/v1/apply/{token}/:
    get:
      operationId: v1_apply_retrieve
      parameters:
      - in: path
        name: token
        schema:
          type: string
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          description: Success
        '404':
          description: Invalid or expired link.
    post:
      operationId: v1_apply_create
      description: |-
        Source 2 — Candidate submits the apply form.

        Identity deduplication:
          1. If a user account already exists with the submitted email/phone →
             return needs_login=True so the frontend shows the login screen.
          2. If a candidate record already exists (no user yet) → append the
             submission data but do NOT create a second candidate row.
          3. If neither exists → create candidate + optional user account.
      parameters:
      - in: path
        name: token
        schema:
          type: string
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          description: Success
        '404':
          description: Invalid or expired link.
...
components:
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT